Calculate distance

+++ domain/src/main/java/xyz/czanik/distanceapp/entities/Location.kt
@@ -1,3 +1,3 @@
 package xyz.czanik.distanceapp.entities
 
-data class Location(val lat: Float, val lng: Float)
+data class Location(val lat: Double, val lng: Double)
